+++ src/main/groovy/com/monits/gradle/sca/StaticCodeAnalysisExtension.groovy
@@ -27,6 +27,7 @@ class StaticCodeAnalysisExtension {
     boolean pmd
     boolean checkstyle
     boolean cpd
+    boolean androidLint
 
     String checkstyleRules
     List<String> pmdRules